Move TS files generated off the OpenAPI spec to a scoped 'src/generated' dir
