+++ .prettierignore
@@ -10,5 +10,5 @@ renovate.json
 
 # Generated
 plugins/bulk-import-backend/api-docs
-plugins/bulk-import-backend/src/openapi.d.ts
-plugins/bulk-import-backend/src/openapidocument.ts
+plugins/bulk-import-backend/src/generated/openapi.d.ts
+plugins/bulk-import-backend/src/generated/openapidocument.ts
