+++ plugins/bulk-import-backend/turbo.json
@@ -9,7 +9,7 @@
       "dependsOn": ["^tsc", "openapi"]
     },
     "openapi": {
-      "outputs": ["src/openapi.d.ts", "src/openapidocument.ts", "api-docs/**"]
+      "outputs": ["src/generated/openapi.d.ts", "src/generated/openapidocument.ts", "api-docs/**"]
     },
     "test": {
       "dependsOn": ["openapi"]
